Module ferro::editor

Structs

Editor

The editor!

Position

A position represented by (x, y) coordinates.

StatusMessage ğŸ”’

A status message printed at the bottom of the editor.

Enums

SearchDirection

The direction in which a search query should be handled.

Constants

DOC_DOWN ğŸ”’

DOC_UP ğŸ”’

LINE_LEFT ğŸ”’

LINE_RIGHT ğŸ”’

PAGE_DOWN ğŸ”’

PAGE_UP ğŸ”’

POS_DOWN ğŸ”’

POS_LEFT ğŸ”’

POS_RIGHT ğŸ”’

POS_UP ğŸ”’

QUIT_TIMES ğŸ”’

VERSION ğŸ”’

WORD_LEFT ğŸ”’

WORD_RIGHT ğŸ”’

Functions

die ğŸ”’