Struct ferro::Row

pub struct Row {
    pub is_highlighted: bool,
    string: String,
    highlighting: Vec<Type>,
    len: usize,
    selections: Vec<[usize; 2]>,
}

Represents a row of text within the document.

Fields
is_highlighted: bool

Whether the row should be highlighted

string: String

The content contained in the row

highlighting: Vec<Type>

The highlight of each grapheme in the row

len: usize

The length of the row’s content

selections: Vec<[usize; 2]>

A list of tuples (start, len) of selections made in the row

Implementations

impl Row

pub fn replace_tabs_with_spaces(&mut self, spaces_per_tab: usize)

Replaces all tabs in the row with spaces.

Arguments

spaces_per_tab - the number of spaces each tab is replaced iwth

pub fn render(&self, start: usize, end: usize) -> String

Renders the row, both the string content of the row and any highlighting.

Arguments

start - the index to start rendering from

end - the index to stop rendering at

pub fn insert(&mut self, at: usize, c: char)

Inserts a character at the given position in the row.

Arguments

at - the position to insert the character at

c - the character to insert

pub fn append(&mut self, other: &Self)

Appends another row to the current row.

Arguments

other - the row to append to this row

pub fn delete(&mut self, at: usize)

Deletes the character at the given position in the row.

Arguments

at - the position of the character in the row to delete

pub fn split(&mut self, at: usize) -> Self

Splits the row into two halves. Afterward, the current row contains the first half while the returned row contains the second half.

Arguments

at - the index in the row to split at

pub fn find(
&self,
query: &str,
at: usize,
direction: SearchDirection
) -> Option<usize>

Finds the index of a string within the row.

Arguments

query - the string to find

at - the index to start finding from

direction - the SearchDirection to use

pub fn find_next_word(
&self,
at: usize,
direction: SearchDirection
) -> Option<usize>

Finds the index of the next word in the row.

Arguments

at - the index to start finding from

direction - the SearchDirection to use

fn find_word_forward(&self, start: usize) -> Option<usize>

Finds the index of the next word in the row in the forward direction.

The index of the word, if found, will be the index immediately preceding the start of that word.

Arguments

start - the starting index of the row

Example

let row = Row::from("Foo Bar");
assert_eq!(row.find_word_forward(1), Some(4));

fn find_word_backward(&self, end: usize) -> Option<usize>

Finds the index of the next word in the row in the backward direction.

The index of the word, if found, will be the index immediately following the end of that word.

Arguments

end - the ending index of the row

Example

let row = Row::from("Foo Bar");
assert_eq!(row.find_word_backward(5), Some(3));

pub fn replace_selections(&mut self, word: &Option<String>)

Replaces all selections made in the row.

Arguments

replace - the string to replace the selections with

pub fn add_selection(&mut self, at: usize, len: usize)

Adds a selection in this row.

Arguments

at - the index of the selection in this row

len - the length of the selection

pub fn reset_selections(&mut self)

Resets all selections made in the row.

pub fn highlight_str(
&mut self,
index: &mut usize,
substring: &str,
chars: &[char],
hl_type: Type
) -> bool

Checks whether there is a string match in the row to highlight.

Arguments

index - the index to check from; this gets updated to the end of the highlight

substring - the string to highlight

chars - the characters in the row

hl_type - the specific highlighting type to use

pub fn highlight_keywords(
&mut self,
index: &mut usize,
chars: &[char],
keywords: &[String],
hl_type: Type
) -> bool

Checks whether there is a keyword to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

chars - the characters in the row

keywords - the keywords that should be highlighted

hl_type - the specific highlighting type to use

fn highlight_primary_keywords(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
chars: &[char]
) -> bool

Checks whether there is a primary keyword to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

chars - the characters in the row

fn highlight_secondary_keywords(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
chars: &[char]
) -> bool

Checks whether there is a secondary keyword to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

chars - the characters in the row

fn highlight_match(&mut self, word: &Option<String>)

Highlights any matches in the row.

Arguments

word - the word to highlight

fn highlight_selection(&mut self)

Highlights any selections in the row.

pub fn highlight_char(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
c: char,
chars: &[char]
) -> bool

Checks whether there is a character literal to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

c - the character at chars[index]

chars - the characters in the row

pub fn highlight_comment(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
_: char,
chars: &[char]
) -> bool

Checks whether there is an inline comment to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

_ - the character at chars[index]

chars - the characters in the row

pub fn highlight_multiline_comment(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
_: char,
chars: &[char]
) -> Option<String>

Checks whether there is a multiline comment to be highlighted in this row. If so, this function returns the delimiter that closes the multiline comment.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

_ - the character at chars[index]

chars - the characters in the row

pub fn highlight_string(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
c: char,
chars: &[char]
) -> bool

Checks whether there is a string literal to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

c - the character at chars[index]

chars - the characters in the row

pub fn highlight_number(
&mut self,
index: &mut usize,
opts: &HighlightingOptions,
c: char,
chars: &[char]
) -> bool

Checks whether there is a number literal to be highlighted.

Arguments

index - the index to check from; this gets updated to the end of the highlight

opts - the HighlightingOptions to use

c - the character at chars[index]

chars - the characters in the row

pub fn highlight(
&mut self,
opts: &HighlightingOptions,
word: &Option<String>,
look_for_multiline_close: &mut Option<String>
)

Computes the highlighting (if any) of every grapheme in this row.

Arguments

opts - the HighlightingOptions to use

word - the word to highlight (if any)

look_for_multiline_close - until this delimiter is found, graphemes are highlighted as a multiline comment. This function changes this accordingly whenever a multiline comment opens or closes.

pub fn len(&self) -> usize

Gets the length of the row.

pub fn is_empty(&self) -> bool

Gets whether this row is empty.

pub fn as_bytes(&self) -> &[u8]

Gets the row’s contents as bytes.

pub fn get_leading_spaces(&self) -> Option<usize>

Gets the number of leading spaces in the row.