Struct ferro::editor::Editor

pub struct Editor {
    should_quit: bool,
    terminal: Terminal,
    cursor_position: Position,
    offset: Position,
    max_position: Option<usize>,
    document: Document,
    status_message: StatusMessage,
    quit_times: u8,
    highlighted_word: Option<String>,
    _sigwinch_flag: Arc<AtomicBool>,
}

The editor!

Fields
should_quit: bool

Whether the editor application should quit

terminal: Terminal

The Terminal that is used

cursor_position: Position

The current position of the cursor

offset: Position

The offset of the visible page

max_position: Option<usize>

The maximal horizontal position that is used when the user navigates up or down

document: Document

The document being edited

status_message: StatusMessage

The status message to be displayed

quit_times: u8

How many more times the quit command needs to be inputted before the application exits

highlighted_word: Option<String>

The word to be highlighted, if any

_sigwinch_flag: Arc<AtomicBool>

Flag for the SIGWINCH signal that is set when the terminal window is resized

Implementations

impl Editor

pub fn default() -> Self

Constructs the default editor.

pub fn run(&mut self)

Runs the editor.

This is essentially an event loop and should only ever be called once.

fn refresh_screen(&mut self) -> Result<(), Error>

Re-renders the terminal screen.

Errors

Will return Err if I/O error encountered

fn draw_status_bar(&self)

Draws the status bar at the bottom of the editor.

fn draw_message_bar(&self)

Draws the message bar at the bottom of the editor.

pub fn draw_row(&self, row: &Row)

Draws a given row on the terminal screen.

Arguments

row - The row to be drawn

fn draw_rows(&self)

fn draw_welcome_message(&self)

Draws a welcome message in the middle of the editor.

fn save(&mut self)

Saves the document being edited.

fn search(&mut self)

Searches for a string in the document.

fn evaluate_expression(&mut self)

Prompts the user for a mathematical expression and displays its evaluated result.

fn process_event(&mut self) -> Result<(), Error>

Processes an event (i.e. a keypress or a mousepress).

Errors

Will return Err if I/O error encountered while reading event

fn process_keypress(&mut self, keypress: Key) -> Result<(), Error>

Processes a keypress event.

Arguments

keypress - the [Key] that was pressed

Errors

Will return Err if I/O error encountered

fn process_mousepress(&mut self, mousepress: MouseEvent) -> Result<(), Error>

Processes a mousepress event.

Arguments

mousepress - the [MouseEvent] that occurred

Errors

Will return Err if I/O error encountered

fn prompt<C>(
&mut self,
prompt: &str,
callback: C
) -> Result<Option<String>, Error>where
C: FnMut(&mut Self, Key, &String),

Prompts the user for input.

Arguments

prompt - the prompt to print

callback - the callback to use

Errors

Will return Err if I/O error encountered

fn prompt_replacement(&mut self) -> Result<Option<String>, Error>

Prompts the user for a string to replace all selections with.

Errors

Will return Err if I/O error encountered

fn scroll(&mut self)

Scrolls the screen by the height of the terminal.

fn move_cursor(&mut self, key: Key)

Moves the cursor based on the key that was pressed.

Arguments

key - The key that was pressed