Struct ferro::Document

pub struct Document {
    pub filename: Option<String>,
    rows: Vec<Row>,
    dirty: bool,
    file_type: FileType,
    spaces_per_tab: usize,
}

The document that is currently being edited.

Fields
filename: Option<String>

The filename of this document

rows: Vec<Row>

The Rows in the document

dirty: bool

Whether the document is dirty

file_type: FileType

The filetype of the document

spaces_per_tab: usize

THe number of spaces that each tab should be replaced with

Implementations

impl Document

pub fn default() -> Self

Constructs a blank document.

pub fn open(filename: &str) -> Result<Self, Error>

Creates a Document from the specified file.

Arguments

filename - the path of the file from which the Document is created

Errors

Will return Err if I/O error encountered while attempting to read file specified by filename

fn calculate_indent(rows: &Vec<Row>) -> usize

Computes the number of spaces for indentation in the file based on a majority algorithm.

Arguments

rows - the Rows to calculate the indent from

fn insert_newline(&mut self, at: &Position) -> usize

Inserts a newline character (‘\n’) at the given position

Arguments

at - the Position to insert the newline character at

pub fn insert(&mut self, at: &mut Position, c: char) -> usize

Inserts a character at the given position

Arguments

at - the Position to insert the character at

c - the character to insert

fn unhighlight_rows(&mut self, start: usize)

pub fn delete(&mut self, at: &Position)

Deletes the character at the given position.

Arguments

at - the Position to delete the character at

pub fn save(&mut self) -> Result<(), Error>

Writes the document to file.

pub fn find(
&self,
query: &str,
at: &Position,
direction: SearchDirection
) -> Option<Position>

Finds the position of the next occurence of a string within the document.

Arguments

query - the string to find

at - the Position to start finding from

direction - the SearchDirection to use

pub fn find_next_word(
&self,
at: &Position,
direction: SearchDirection
) -> Option<Position>

Finds the position of the next word in the document.

A word is defined as a sequence of alphanumeric characters.

Arguments

at - the position to start looking from

direction - the SearchDirection to use

pub fn highlight(&mut self, word: &Option<String>, until: Option<usize>)

Computes the highlight of all rows in the document.

Arguments

word - the word to highlight, if any

until - the index to stop highlighting at

pub fn refresh_highlighting(&mut self)

Re-computes all highlighting.

pub fn add_selection(&mut self, at: Position, len: usize)

Adds a selection in the document.

Arguments

at - the position of the selection within the document

len - the length of the selection

pub fn delete_selections(&mut self)

Deletes all selections made in the document.

pub fn replace_selections(&mut self, replace: &Option<String>)

Replaces all selections made in the document.

Arguments

replace - the string to replace the selections with

pub fn reset_selections(&mut self)

Resets all selections made in the document.

pub fn row(&self, index: usize) -> Option<&Row>

Gets a row in the document.

Arguments

index - the row’s index

pub fn len(&self) -> usize

Gets the number of rows in the document.

pub fn is_empty(&self) -> bool

Gets whether the document is entirely empty.

pub fn is_dirty(&self) -> bool

Gets whether the document is dirty.

pub fn file_type(&self) -> String

Gets the document’s filetype.